# mvcc: reuse iter for get and put in cput