@tamird 2016-02-10
LGTM. Nice.

---

Reviewed 2 of 2 files at r1.
Review status: all files reviewed at latest revision, 2 unresolved discussions.

---

*[storage/engine/mvcc.go, line 786 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACO1f8zC8rHc4lqtkl:-KACO1f8zC8rHc4lqtkm:-1280936503) ([raw file](https://github.com/cockroachdb/cockroach/blob/97e7ae5eef6c46cb216aa23ec885825d4ca8da35/storage/engine/mvcc.go#L786)):*
remove

---

*[storage/engine/mvcc.go, line 789 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACO2iJYCYArG7Bt9Qj:-KACO2iJYCYArG7Bt9Qk:-1276068445) ([raw file](https://github.com/cockroachdb/cockroach/blob/97e7ae5eef6c46cb216aa23ec885825d4ca8da35/storage/engine/mvcc.go#L789)):*
remove (or add a similar newline in `MVCCGet`

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-:-KACOEN6ZneBS_xf5O_y:1324449798)*
<!-- Sent from Reviewable.io -->