@dt 2016-02-10
Review status: 1 of 2 files reviewed at latest revision, 2 unresolved discussions.

---

*[storage/engine/mvcc.go, line 786 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACO1f8zC8rHc4lqtkl:-KACQdm6HAPPLUVGuMdw:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/97e7ae5eef6c46cb216aa23ec885825d4ca8da35/storage/engine/mvcc.go#L786)):*
Done.

---

*[storage/engine/mvcc.go, line 789 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACO2iJYCYArG7Bt9Qj:-KACQdSHcd2T3LJYaUMK:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/97e7ae5eef6c46cb216aa23ec885825d4ca8da35/storage/engine/mvcc.go#L789)):*
Done.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308)*
<!-- Sent from Reviewable.io -->