@dt 2016-02-10
I'm still not sure if it is worth merging -- it doesn't seems to move the SQL benchmarks appreciably.

---

Review status: 1 of 2 files reviewed at latest revision, 2 unresolved discussions.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-:-KACR6Zl5C9rcnlJx8oD:115916058)*
<!-- Sent from Reviewable.io -->