@tschottdorf 2016-02-10
The performance doesn't really indicate that we must merge this, but it seems "pretty" to not run through more than one iterator for one "operation". @petermattis?