@tamird 2016-02-10
Yeah, this is a nice cleanup as far as I'm concerned.