@petermattis 2016-02-11
LGTM. I was hoping this was going to have a bigger impact. Regardless, I think this is worth merging. 

---

Review status: all files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---

*[storage/engine/mvcc.go, line 544 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACmMA5KPbFitnSme0Q:-KACmMA5KPbFitnSme0R:1084846760) ([raw file](https://github.com/cockroachdb/cockroach/blob/4e2f838e3efe6c5b5f499a7a21ec2aad95b2fab1/storage/engine/mvcc.go#L544)):*
I don't think the `engine` parameter is used anymore. For a bit of additional consistency with other functions in this file (e.g. `mvccGetInternal`), I'd make `iter` the first parameter. 

---

*[storage/engine/mvcc.go, line 794 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACmnrAsuMtYwF7TQbn:-KACmnrAsuMtYwF7TQbo:1225969468) ([raw file](https://github.com/cockroachdb/cockroach/blob/4e2f838e3efe6c5b5f499a7a21ec2aad95b2fab1/storage/engine/mvcc.go#L794)):*
Let's make `iter` the second parameter after `engine`.

---

*[storage/engine/mvcc.go, line 1019 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACn0gEfCiC1S81P220:-KACn0gEfCiC1S81P221:394453950) ([raw file](https://github.com/cockroachdb/cockroach/blob/4e2f838e3efe6c5b5f499a7a21ec2aad95b2fab1/storage/engine/mvcc.go#L1019)):*
One additional change to be made: when `mvccPutUsingIter` is called here, the `iter` might already be pointing at the correct key. This will show up in the `mvccGetMetadata` call done at the beginning of `mvccPutInternal`. I think we can do a quick check at the beginning of `mvccGetMetadata` and avoid performing a call to seek. 

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-:-KACnMCHRRvGlfWW17G4:2110393007)*
<!-- Sent from Reviewable.io -->