@dt 2016-02-11
Review status: 1 of 2 files reviewed at latest revision, 3 unresolved discussions.

---

*[storage/engine/mvcc.go, line 544 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACmMA5KPbFitnSme0Q:-KADYZYjsI35RgGECE6a:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/4e2f838e3efe6c5b5f499a7a21ec2aad95b2fab1/storage/engine/mvcc.go#L544)):*
Done.

---

*[storage/engine/mvcc.go, line 794 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACmnrAsuMtYwF7TQbn:-KADYZDRrVdY88dQd9ST:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/4e2f838e3efe6c5b5f499a7a21ec2aad95b2fab1/storage/engine/mvcc.go#L794)):*
Done.

---

*[storage/engine/mvcc.go, line 1019 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACn0gEfCiC1S81P220:-KADYbOWLzmpJBvLZWa3:1989975871) ([raw file](https://github.com/cockroachdb/cockroach/blob/4e2f838e3efe6c5b5f499a7a21ec2aad95b2fab1/storage/engine/mvcc.go#L1019)):*
Hm, my first attempt at that didn't quite work (compiles but isn't correct). Either way, I'm inclined to split do that in a followup.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308)*
<!-- Sent from Reviewable.io -->