@petermattis 2016-02-11
Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---

*[storage/engine/mvcc.go, line 1019 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-KACn0gEfCiC1S81P220:-KAFnGBS6SBxco-2s0Zv:-1172894286) ([raw file](https://github.com/cockroachdb/cockroach/blob/4e2f838e3efe6c5b5f499a7a21ec2aad95b2fab1/storage/engine/mvcc.go#L1019)):*
A follow-on PR is fine by me.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308)*
<!-- Sent from Reviewable.io -->