@petermattis 2016-02-11
LGTM

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4308#-:-KAFnL71XA_U4_0j-27m:676834233)*
<!-- Sent from Reviewable.io -->